+++ src/main/java/io/stargate/sgv2/jsonapi/api/model/command/clause/filter/ComparisonExpression.java
@@ -17,7 +17,7 @@ import org.bson.types.ObjectId;
  * {"username" : {"$eq" : "aaron"}} In here we expand the shortcut into a canonical long form, so it
  * is all the same.
  */
-public class ComparisonExpression implements Invertible {
+public class ComparisonExpression {
 
   private final String path;
 
@@ -33,12 +33,8 @@ public class ComparisonExpression implements Invertible {
     return filterOperations;
   }
 
-  /**
-   * implements Invertible, method to invert a ComparisonExpression this method will be called when
-   * $not operator is pushed down
-   */
-  @Override
-  public Object invert() {
+  /** This method will be called when not operation is pushed down */
+  public void flip() {
     List<FilterOperation<?>> filterOperations = new ArrayList<>(this.filterOperations.size());
     for (FilterOperation<?> filterOperation : this.filterOperations) {
       final FilterOperator flippedOperator = filterOperation.operator().flip();
@@ -47,7 +43,6 @@ public class ComparisonExpression implements Invertible {
       filterOperations.add(new ValueComparisonOperation<>(flippedOperator, operand));
     }
     this.filterOperations = filterOperations;
-    return this;
   }
 
   /**
