--- src/main/java/io/stargate/sgv2/jsonapi/api/model/command/clause/filter/Invertible.java
@@ -1,8 +1,0 @@
-package io.stargate.sgv2.jsonapi.api.model.command.clause.filter;
-
-/** Interface Invertible implemented by comparisonExpression and logicalExpression */
-public interface Invertible {
-
-  /** invert method to invert a comparisonExpression or logicalExpression */
-  Object invert();
-}
