+++ src/main/java/io/stargate/sgv2/jsonapi/api/model/command/deserializers/FilterClauseDeserializer.java
@@ -44,7 +44,6 @@ public class FilterClauseDeserializer extends StdDeserializer<FilterClause> {
   @Override
   public FilterClause deserialize(
       JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
-
     JsonNode filterNode = deserializationContext.readTree(jsonParser);
     if (!filterNode.isObject()) {
       throw ErrorCode.UNSUPPORTED_FILTER_DATA_TYPE.toApiException();
@@ -53,6 +52,8 @@ public class FilterClauseDeserializer extends StdDeserializer<FilterClause> {
     LogicalExpression implicitAnd = LogicalExpression.and();
     populateExpression(implicitAnd, filterNode);
     validate(implicitAnd);
+    // push down not operator
+    implicitAnd.traverseForNot(null);
 
     return new FilterClause(implicitAnd);
   }
